refactor(hickey): hide RenderCtx data constructor; force smart-ctor path
Per /hickey: now that the record gains a userData field, future fields
will keep silently breaking any external positional construction site.
The smart constructors (`mkRenderCtx`, `emptyRenderCtx`) exist precisely
to absorb such extensions — they should be the only construction path.

Move the data declaration to a new `Heist.Extra.Splices.Pandoc.Ctx.Internal`
module that exposes the constructor for in-package consumers (Render.hs
uses RecordWildCards). The public `Ctx` and umbrella `Pandoc` modules
re-export the type and field accessors but not the constructor, so
downstream callers must go through the smart constructors.

Field-update syntax (`ctx { blockSplice = ... }`) and field-accessor
reads continue to work since accessors stay exported.
