+++ heist-extra.cabal
@@ -98,6 +98,7 @@ library
     Heist.Extra.Splices.Pandoc
     Heist.Extra.Splices.Pandoc.Attr
     Heist.Extra.Splices.Pandoc.Ctx
+    Heist.Extra.Splices.Pandoc.Ctx.Internal
     Heist.Extra.Splices.Pandoc.Footnotes
     Heist.Extra.Splices.Pandoc.Render
     Heist.Extra.Splices.Pandoc.Render.Internal
